scanner: Move common test code to AbstractScannerTest
